+++ hadoop-hdds/common/src/main/java/org/apache/hadoop/hdds/utils/BackgroundService.java
@@ -50,6 +50,7 @@ public abstract class BackgroundService {
   private final int threadPoolSize;
   private final String threadNamePrefix;
   private final PeriodicalTask service;
+  private CompletableFuture<Void> future;
 
   public BackgroundService(String serviceName, long interval,
       TimeUnit unit, int threadPoolSize, long serviceTimeout) {
@@ -68,6 +69,11 @@ public abstract class BackgroundService {
     this.threadNamePrefix = threadNamePrefix;
     initExecutorAndThreadGroup();
     service = new PeriodicalTask();
+    this.future = CompletableFuture.completedFuture(null);
+  }
+
+  protected CompletableFuture<Void> getFuture() {
+    return future;
   }
 
   @VisibleForTesting
@@ -138,7 +144,7 @@ public abstract class BackgroundService {
 
       while (!tasks.isEmpty()) {
         BackgroundTask task = tasks.poll();
-        CompletableFuture.runAsync(() -> {
+        future = future.thenCombine(CompletableFuture.runAsync(() -> {
           long startTime = System.nanoTime();
           try {
             BackgroundTaskResult result = task.call();
@@ -157,7 +163,7 @@ public abstract class BackgroundService {
                   serviceName, endTime - startTime, serviceTimeoutInNanos);
             }
           }
-        }, exec);
+        }, exec), (Void1, Void) -> null);
       }
     }
   }
